Move command handler stuff to a separate file
